FY23Q4 refresh - msteams-conversation-bots
- run through all exercise steps in associated MSLearn module to validate good working order
- updated steps & fixed typos where necessary
- updated slides where necessary
- update code samples to reflect MSLearn HOL exercise refresh & updates
- any & all users/names/company names/pictures/IDs are not real people/tenants... all are from M365 demo tenants provided by Microsoft
