+++ Teams/40 Conversational Bots/Demos/01-learn-msteams-bots/package.json
@@ -23,8 +23,7 @@
     "express-session": "~1.17.2",
     "morgan": "~1.10.0",
     "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "msteams-react-base-component": "^4.0.1"
+    "react-dom": "^16.8.6"
   },
   "devDependencies": {
     "@types/debug": "4.1.7",
@@ -44,8 +43,8 @@
     "vinyl": "2.2.1",
     "webpack": "^5.72.0",
     "yargs": "^17.5.1",
-    "yoteams-build-core": "^1.7.0",
-    "yoteams-deploy": "^1.2.0",
+    "yoteams-build-core": "^1.8.0",
+    "yoteams-deploy": "^1.4.0",
     "webpack-node-externals": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^5.22.0",
     "@typescript-eslint/parser": "^5.22.0",
