+++ Teams/40 Conversational Bots/Demos/01-learn-msteams-bots/src/server/TeamsAppsComponents.ts
@@ -1,4 +1,4 @@
 // Components will be added here
 export const nonce = {}; // Do not remove!
-// Automatically added for the conversationalBot bot
-export * from "./conversationalBot/ConversationalBot";
+// Automatically added for the conversationalBotBot bot
+export * from "./conversationalBotBot/ConversationalBotBot";
