+++ Teams/40 Conversational Bots/Demos/01-learn-msteams-bots/src/server/conversationalBotBot/dialogs/helpDialog.ts
