+++ Teams/40 Conversational Bots/Demos/02-learn-msteams-bots/.yo-rc.json
@@ -4,6 +4,6 @@
       "developer": "Contoso"
     },
     "libraryName": "conversationalBot",
-    "generator-version": "4.0.1"
+    "generator-version": "4.1.0"
   }
 }
