+++ Teams/40 Conversational Bots/Demos/02-learn-msteams-bots/src/server/conversationalBotBot/cards/welcomeCard.ts
