--- Teams/40 Conversational Bots/Demos/03-learn-msteams-bots/.env.example
@@ -1,34 +1,0 @@
-# The public domain name of where you host your application
-PUBLIC_HOSTNAME=conversationalbot.azurewebsites.net
-
-# Id of the Microsoft Teams application
-APPLICATION_ID=
-# Package name of the Microsoft Teams application
-PACKAGE_NAME=
-
-# App Id and App Password ofr the Bot Framework bot
-MICROSOFT_APP_ID=
-MICROSOFT_APP_PASSWORD=
-
-# Port for local debugging
-PORT=3007
-
-# Security token for the default outgoing webhook
-SECURITY_TOKEN=
-
-# ID of the Outlook Connector
-CONNECTOR_ID=
-
-# Application Insights instrumentation key
-APPINSIGHTS_INSTRUMENTATIONKEY=
-
-# NGROK configuration for development
-# NGROK authentication token (leave empty for anonymous)
-NGROK_AUTH=
-# NGROK sub domain. ex "myapp" or  (leave empty for random)
-NGROK_SUBDOMAIN=
-# NGROK region. (us, eu, au, ap - default is us)
-NGROK_REGION=
-
-# Debug settings, default logging "msteams"
-DEBUG=msteams
