+++ Teams/40 Conversational Bots/Demos/03-learn-msteams-bots/package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "@fluentui/react-northstar": "~0.62.0",
     "@microsoft/teams-js": "^2.0.0",
+    "adaptive-expressions": "4.19.3",
+    "adaptivecards-templating": "2.3.1",
     "axios": "^0.27.2",
     "botbuilder": "~4.16.0",
     "botbuilder-dialogs": "~4.16.0",
@@ -23,8 +25,7 @@
     "express-session": "~1.17.2",
     "morgan": "~1.10.0",
     "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "msteams-react-base-component": "^4.0.1"
+    "react-dom": "^16.8.6"
   },
   "devDependencies": {
     "@types/debug": "4.1.7",
@@ -33,7 +34,17 @@
     "@types/morgan": "~1.9.3",
     "@types/react": "16.8.10",
     "@types/react-dom": "^16.9.7",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
     "browser-sync": "^2.26.5",
+    "eslint": "^8.15.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-webpack-plugin": "^3.0.1",
     "fancy-log": "^2.0.0",
     "file-loader": "~6.2.0",
     "fork-ts-checker-webpack-plugin": "~7.2.11",
@@ -43,20 +54,10 @@
     "typestyle": "~2.3.0",
     "vinyl": "2.2.1",
     "webpack": "^5.72.0",
-    "yargs": "^17.5.1",
-    "yoteams-build-core": "^1.7.0",
-    "yoteams-deploy": "^1.2.0",
     "webpack-node-externals": "^3.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.22.0",
-    "@typescript-eslint/parser": "^5.22.0",
-    "eslint": "^8.15.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-webpack-plugin": "^3.0.1"
+    "yargs": "^17.5.1",
+    "yoteams-build-core": "^1.8.0",
+    "yoteams-deploy": "^1.4.0"
   },
   "browserslist": [
     "> 1%",
