+++ Teams/40 Conversational Bots/Demos/03-learn-msteams-bots/src/server/conversationalBotBot/cards/welcomeCard.json
